# Improve pure Go speed up to 45%