* Only accumulate once to memory.
* Use cast instead of AND to truncate.
* Use binary.LittleEndian - observed some cases where it didn't use single instruction to load while testing.
* Make code a bit easier to copy+paste. Tried a function, but it wouldn't inline properly.

```
benchmark                                   old ns/op     new ns/op     delta
BenchmarkHasher64/16/go/plain-32            10.8          10.8          +0.18%
BenchmarkHasher64/16/go/seed-32             11.5          11.5          -0.52%
BenchmarkHasher64/64/go/plain-32            13.2          13.5          +1.89%
BenchmarkHasher64/64/go/seed-32             14.5          14.5          -0.48%
BenchmarkHasher64/256/go/plain-32           42.0          38.4          -8.48%
BenchmarkHasher64/256/go/seed-32            55.3          53.3          -3.74%
BenchmarkHasher64/1024/go/plain-32          116           96.8          -16.17%
BenchmarkHasher64/1024/go/seed-32           131           113           -13.91%
BenchmarkHasher64/4096/go/plain-32          411           320           -22.10%
BenchmarkHasher64/4096/go/seed-32           408           324           -20.58%
BenchmarkHasher64/16384/go/plain-32         1575          1220          -22.54%
BenchmarkHasher64/16384/go/seed-32          1567          1245          -20.55%
BenchmarkHasher64/65536/go/plain-32         6215          4825          -22.37%
BenchmarkHasher64/65536/go/seed-32          6170          4900          -20.58%
BenchmarkHasher64/262144/go/plain-32        24618         19150         -22.21%
BenchmarkHasher64/262144/go/seed-32         24556         19472         -20.70%
BenchmarkHasher64/1048576/go/plain-32       100340        76903         -23.36%
BenchmarkHasher64/1048576/go/seed-32        98686         78452         -20.50%
BenchmarkHasher64/4194304/go/plain-32       397258        310424        -21.86%
BenchmarkHasher64/4194304/go/seed-32        393518        316373        -19.60%
BenchmarkHasher64/16777216/go/plain-32      1759263       1392247       -20.86%
BenchmarkHasher64/16777216/go/seed-32       1795954       1420991       -20.88%
BenchmarkHasher64/67108864/go/plain-32      8285671       5713840       -31.04%
BenchmarkHasher64/67108864/go/seed-32       8034171       5773070       -28.14%
BenchmarkHasher64/268435456/go/plain-32     33112283      22882546      -30.89%
BenchmarkHasher64/268435456/go/seed-32      33193939      23083580      -30.46%

benchmark                                   old MB/s     new MB/s     speedup
BenchmarkHasher64/16/go/plain-32            1479.39      1475.91      1.00x
BenchmarkHasher64/16/go/seed-32             1386.08      1393.51      1.01x
BenchmarkHasher64/64/go/plain-32            4836.06      4748.53      0.98x
BenchmarkHasher64/64/go/seed-32             4401.57      4422.27      1.00x
BenchmarkHasher64/256/go/plain-32           6100.65      6667.03      1.09x
BenchmarkHasher64/256/go/seed-32            4625.77      4805.92      1.04x
BenchmarkHasher64/1024/go/plain-32          8866.30      10576.22     1.19x
BenchmarkHasher64/1024/go/seed-32           7829.90      9094.16      1.16x
BenchmarkHasher64/4096/go/plain-32          9967.92      12796.39     1.28x
BenchmarkHasher64/4096/go/seed-32           10035.09     12632.76     1.26x
BenchmarkHasher64/16384/go/plain-32         10399.26     13433.97     1.29x
BenchmarkHasher64/16384/go/seed-32          10455.46     13160.54     1.26x
BenchmarkHasher64/65536/go/plain-32         10544.91     13583.15     1.29x
BenchmarkHasher64/65536/go/seed-32          10621.16     13373.47     1.26x
BenchmarkHasher64/262144/go/plain-32        10648.57     13688.70     1.29x
BenchmarkHasher64/262144/go/seed-32         10675.56     13462.41     1.26x
BenchmarkHasher64/1048576/go/plain-32       10450.25     13634.96     1.30x
BenchmarkHasher64/1048576/go/seed-32        10625.35     13365.79     1.26x
BenchmarkHasher64/4194304/go/plain-32       10558.14     13511.52     1.28x
BenchmarkHasher64/4194304/go/seed-32        10658.47     13257.47     1.24x
BenchmarkHasher64/16777216/go/plain-32      9536.50      12050.46     1.26x
BenchmarkHasher64/16777216/go/seed-32       9341.67      11806.70     1.26x
BenchmarkHasher64/67108864/go/plain-32      8099.39      11744.97     1.45x
BenchmarkHasher64/67108864/go/seed-32       8352.93      11624.47     1.39x
BenchmarkHasher64/268435456/go/plain-32     8106.82      11731.01     1.45x
BenchmarkHasher64/268435456/go/seed-32      8086.88      11628.85     1.44x
```